services/RAG_API/requirements.txt
# ============================================
# RAG API Dependencies
# FILE: services/RAG_API/requirements.txt
# ============================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9

# Azure Services
azure-search-documents==11.6.0
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.20.0

# OpenAI
openai==1.44.0

# Data Processing
pandas==2.2.2
numpy==1.24.4
pyarrow==16.1.0
duckdb==1.0.0

# Text Processing
tiktoken==0.7.0

# Environment
python-dotenv==1.0.1

# Utilities
requests==2.31.0